requirements.txt
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
tqdm==4.62.3
visitor==0.1.3
Werkzeug==2.0.2
